// app/lib/types/audit.db.types.ts
export enum AuditAction {
  CREATE = "CREATE",
  UPDATE = "UPDATE",
  DELETE = "DELETE",
  RESTORE = "RESTORE",
  IMAGE_ADD = "IMAGE_ADD",
  IMAGE_REMOVE = "IMAGE_REMOVE",
  STATUS_CHANGE = "STATUS_CHANGE",
  INVENTORY_RESERVATION = "INVENTORY_RESERVATION",
  INVENTORY_RESERVATION_PARTIAL = "INVENTORY_RESERVATION_PARTIAL",
  INVENTORY_RESERVATION_CANCEL = "INVENTORY_RESERVATION_CANCEL",
  CHECKOUT_FAILURE = "CHECKOUT_FAILURE",
  CHECKOUT_SUCCESS = "CHECKOUT_SUCCESS",
  PAYMENT_FAILURE = "PAYMENT_FAILURE",
  PAYMENT_SUCCESS = "PAYMENT_SUCCESS",
  REFUND = "REFUND",
  ORDER_CREATED = "ORDER_CREATED",
  ORDER_UPDATED = "ORDER_UPDATED",
  ORDER_DELETED = "ORDER_DELETED",
  ORDER_FAILURE = "ORDER_FAILURE",
  ORDER_STATUS_CHANGE = "ORDER_STATUS_CHANGE",
}
export enum AuditSource {
  WEB = "WEB",
  API = "API",
  SYSTEM = "SYSTEM",
}
export enum EntityType {
  PRODUCT = "PRODUCT",
  USER = "USER",
  ORDER = "ORDER",
}
// audit.db.types.ts
export interface AuditLogDetails {
  success: boolean;
  device: DeviceInfo;
  message: string;
  // [key: string]: any; // Additional dynamic details
}
export interface ClientAuditLogDetails {
  details: AuditLogDetails;
  action: SecurityAuditAction;
  timestamp: Date;
  // message: string;
}
export interface IGAuditLogDB {
  _id: string;
  actor: string;
  action: AuditAction;
  entity_type: EntityType;
  entity_id: String;
  //   ipAddress?: string;
  //   userAgent?: string;
  source: AuditSource;
  correlation_id: string;
  context?: Record<string, any>;
  created_at: Date;
  updated_at: Date;
}
export interface IAuditLogChangesDB {
  _id: string;
  audit_log_id: string;
  field: string;
  before: any;
  after: any;
  change_type: "ADD" | "MODIFY" | "REMOVE";
  created_at: Date;
}